Add ability to add your own pytest flags to `make test`
Previously, the `make test` command would just run `python -m pytest`
directly, without an option to add pytest flags.

One thing that I find useful is to run `python -m pytest --pdb` so that
any test that fails goes straight into breakpoint mode on the line that
raised the error.

Though this is useful for me, it isn't necessarily useful for anyone
else, thus adding a `PYTEST_FLAGS` (in the same vain as `CFLAGS` in C
projects), will allow for developers to pick and choose whichever flags
they want to run tests with.