Apply django_stubs to example_app used for testing
Prior to this commit, there was a rogue type error on the `models.py`
file under example_app. This was because we weren't running django-stubs
to make sure the example_app was fully typed.

This commit changes that, pointing django-stubs towards the settings of
that app makes the type error "disappear", i.e., django-stubs adds the
proper annotations.