Do not migrate if the router check does not pass
Prior to this change, we weren't checking if the database router allowed
the migration to proceed.

We thus add the router.allow_migrate() check, which is the same that the
RunSQL operation runs internally. Ref:

- https://github.com/django/django/blob/7adb6dd98d50a238f3eca8c15b16b5aec12575fd/django/db/migrations/operations/special.py#L105

A possible error that would raise before this fix happens when the
Django applicatoin has multiple databases and the migration should be
routed to only apply on a particular database.