Split SaferAddIndexConcurrently basic logic into a base class
In future commits, we want to have a new class called
SaferRemoveIndexConcurrently, which will basically be the same as
SaferAddIndexConcurrently but with opposite behaviour.

These functions that are now being moved to the base class will be
useful in this future SaferRemoveIndexConcurrently class and are thus
extracted in this commit.

The diff might look a bit strange, but essentially this commit moves the
database_forwards/backwards functions out (along with the auxiliary
functions that make it work) into the new BaseIndexOperation class,
and calls it from the existing SaferAddIndexConcurrently class.