Make SaferAddIndexConcurrently inherit from AddIndexConcurrently
Not using RunSQL as a base means that I can simplify much of the code in
this class. Plus, I can also make the BaseIndexOperation more generic so
that it can be easily used by SaferRemoveIndexConcurrently

Note that the `hint` argument isn't available on Django's original
AddIndexConcurrently class, and it is not necessary any longer as we
don't use the router but the model itself to check if the migration can
happen. Thus, the `hint` argument has been removed from the
implementation and documentation.