Add MigrationTimeoutOptions dataclass
This new dataclass is responsible for capturing the options necessary to
run a migration with timeouts.

It extracts the logic out of the Command class, so that once we add
further configuration (a retry mechanism), the Command class can remain
the same, at the expense of this dataclass growing instead.