Make BaseIndexOperation inherit basic functionality
Instead of passing the operations around, we can just make the
BaseIndexOperation class inherit the behaviour we need.

This will be useful as we add a new operation (SaferAddUniqueConstraint)
that doesn't inherit from the Index operation class from Django itself
but, in our case, needs to be able to trigger indexes concurrently and
thus not be in a transaction.