Add time since the migration started in RetryState
It's useful to know how long a particular migration has been retrying
for.

This commit will expose this information to the RetryState object that
is passed to retry callbacks.

This will allow these callbacks to do interesting things such as logging
how much time has elapsed since the migration started.