Pop stdout before passing it down to the command
If the `stdout` is not popped, the call on line 114 will fail:

  super().handle(*args, stdout=stdout, **options)

This happens because `stdout` is manually passed and the `options`
variable *also* has an entry with the "stdout" key.

This effectively means this call does:

  super().handle(..., stdout=stdout, stdout=options["stdout"], ...)

Along with the fix, the test `test_when_stdout_is_passed_in` has been
adapted to test the expected behaviour instead of showcasing the error.