Remove dependency on instance attributes
The function _ensure_no_lock_timeout_set sets an instance variable
called original_lock_timeout so that the timeout can be reset at the
bottom of the operation.

We don't want to rely on instance variables that are created on the fly
as we want to stop using this class in inheritance mode.

This refactor replaces the functions using the instance variable to two
simpler functions that just show and set the lock_timeout postgres
setting.