Fix broken test in different Django versions
Sadly Django changed the __init__ interface. For example in 5.0 we have:

```
class CheckConstraint(BaseConstraint):
    def __init__(
        self, *, check, name, violation_error_code=None, violation_error_message=None
    ):
```

Whereas in 5.1 we have:

```
class CheckConstraint(BaseConstraint):
    # RemovedInDjango60Warning: when the deprecation ends, replace with
    # def __init__(
    #  self, *, condition, name, violation_error_code=None, violation_error_message=None
    # )
    def __init__(
        self,
        *,
        name,
        condition=None,
        check=None,
        violation_error_code=None,
        violation_error_message=None,
    ):
```

We could just use "check" instead, but as the comment says, it will be
deprecated soon.

As we don't care which class this is called with, only that it must be a
UniqueConstraint, we just place a BaseConstraint here for the test to be
fixed.