Introduce IndexSQLBuilder
At present, the SafeIndexOperationManager class relies on a models.Index
instance being passed on to generate the CREATE INDEX CONCURRENTLY ddl.

The models.Index.create_sql function will error if the state of the
model does not match the index sql that would be created.

This will become a problem when a future operation to create FKs comes
about. That operation will not entirely rely on the model state, as it
needs to perform at least 4 steps before the state is updated, one of
which is the (optional) creation of an index concurrently (for the fk
field).

With this class, we only need to take some metadata about the model
(name, table name, and column name) and we can proceed to generate a SQL
statement directly.