Make introspection queries not break on sqlmigrate
Many of the introspection queries here break when the `sqlmigrate`
command runs without an up-to-date database schema.

For example, take this introspection query for checking if the column
already exists:

  SELECT 1
  FROM pg_catalog.pg_attribute
  WHERE
      attrelid = {table_name}::regclass
      AND attname = {column_name};

If this introspection query run for real, in a database where the
{table_name} does not exist, it would return the following:

  django.db.utils.ProgrammingError: relation "table_name" does not exist
  LINE 5:     attrelid = 'table_name::regclass'

This happens because in certain cases the database doesn't have all
migrations applied and thus doesn't have a schema computed when the
sqlmigrate command runs.

This commit adds a workaround, by not actually running the introspection
queries when sqlmigrate is running.

The introspection queries are also not printed to avoid confusion, as
they are not part of the plan per-se, and are more of a way for us to
make sure the operation is reentrant and idempotent.