Run migrate_with_timeouts with an advisory lock
Prior to this change, the migrate_with_timeouts command could run in
parallel multiple times.

This is a problem because many migration operations that are multi-step
and non-atomic will behave badly under such circumstances.

To prevent this of happening, this command will now require a
session-level advisory lock for the entirety of its procedure.

The session-level advisory lock will be automatically removed once the
management command finishes, whether successfully or not.

See:
  https://github.com/django/django/blob/857b1048/django/core/management/base.py#L426