Introduce SaferAddFieldOneToOne operation
The default SQL DDL statements produced by Django to create a
OneToOneField are not safe to run and can cause outages.

For example, when the following foo field is added to Bar:

  class Bar(models.Model):
      foo = models.OneToOneField(
          Foo,
          on_delete=models.CASCADE,
          primary_key=False,  # False is the default
          null=True,
      )

A new migration is created via makemigrations with the following
operation:

  operations = [
      migrations.AddField(
          model_name='bar',
          name='foo',
          field=models.OneToOneField(
              null=True,
              on_delete=django.db.models.deletion.CASCADE,
              to='blog.foo'
          ),
      ),
  ]

This operation, in turn, produces the following statements:

  BEGIN;
  --
  -- Add field foo to bar
  --
  ALTER TABLE "blog_bar"
  ADD COLUMN "foo_id" bigint NULL
  UNIQUE CONSTRAINT "auto_gen_constraint_name"
  REFERENCES "blog_foo"("id")
  DEFERRABLE INITIALLY DEFERRED;

  SET CONSTRAINTS "auto_gen_constraint_name" IMMEDIATE;
  COMMIT;

This is not safe, the ALTER TABLE requires an AccessExclusiveLock on
the bar table, and a ShareRowExclusiveLock on the foo table. Same for
the SET CONSTRAINTS command.

The new SaferAddFieldOneToOne provides an alternative, by creating both
the unique constraint and the fk constraint using the safer operations
that are already included in this library.