Add a Main Features section in the README.md
Prior to this change, the README.md file was looking rather empty. The
only important information it contained was a link to the full
documentation (ReadTheDocs page).

In some cases, potential users might come to the GitHub repository
page directly and face a blank wall.

This commit adds a summary of the main features to the README.md with
some click links for interested parties to navigate them.