Fix database_backwards arguments order
The new_state (from_state) should come before the project_state
(to_state), as this is a backwards operation.

There are no changes with regards to how the package works, this is
only about test set up.

This change revealed a few problems with the set-up of some tests which
have been corrected so that this change could go through.

This set-up has been copied from Django's tests, for example:

https://github.com/django/django/blob/51df0dff3c4f28016185a9e876ee5b3420712f99/tests/migrations/test_operations.py#L308-L314

But something along the way (potentially a wrong ctrl+c/ctrl+v) meant
that the order of backward operations remained wrong.