Introduce SaferRenameModelPart1
Modern applications usually run schema changes before they deploy new
code to production. For example, in blue/green deployment, there is a
gradual transfer of traffic from the servers using the old code (blue),
to servers using the new code (green).

Once all traffic is migrated to the "green" servers, the "blue" servers
usually remain ready in case the application needs to be rolled back.

In this context, a Python/Django application that just performed a
RenameModel operation, the "blue" servers will still hold the old Django
state in which the table is believed to have the old name. Only the
"green" servers will have the correct Django state with the correct new
table name. This means that any queries performed by the "blue" server
against the model being renamed will fail.

Analogously, if the migration has to be reverted and the traffic
gradually moved from green to blue, the green servers would in turn have
the wrong Django state.

This new class performs the following SQL statements to rename a table
in a safer way than Django's default operation (RenameModel)

  BEGIN;

  ALTER TABLE foo RENAME TO new_foo;

  CREATE VIEW foo AS SELECT * FROM new_foo;

  COMMIT;

This is the first step (out of two) to rename a table in a situation
where both old code (before the change) and new code (after the change)
might be running at the same time.

The ALTER TABLE acquires a quick AccessExclusive lock, and the CREATE
VIEW AS SELECT operation is also very quick, as it only creates a "view
alias" to the renamed table. This view alias accepts all operations
(INSERT, UPDATE, DELETE, SELECT). The docs provide further explanation:

> Simple views are automatically updatable: the system will allow
> INSERT, UPDATE and DELETE statements to be used on the view in the same
> way as on a regular table. A view is automatically updatable if it
> satisfies all of the following conditions: (...) - All columns in the
> view's select list must be simple references to columns of the
> underlying relation. They cannot be expressions, literals or functions.
> System columns cannot be referenced, either.

It doesn't matter if the table:

- has a trigger
- has a constraint

The view alias will work properly.

Note that this operation does not update the name of other relations
that have been created using the old table name, for example Foreign Key
constraints, M2M tables, etc. Django actually drops constraints and
recreates them from scratch during the RenameModel operation, which is
another reason to not use Django's RenameModel.