Introduce SaferRenameModelPart2
This is the second and final step for performing a safer RenameModel
operation.

Essentially, this operation is clean-up only. It allows for removing the
view created by step 1 once the user has attested that no more "blue"
servers are using the view.

It essentially performs a:

  DROP VIEW IF EXISTS <view_name>;

There is one important limitation of this operation. Namely, the old/new
states are different from the Part 1 operation since Part 2 will run in a
different migration. This means that we lose access to `old_model` and
only have access to the `new_model`.

Without access to the old model, we don't know what the old table name
was. In this regard, the old_table_name parameter is required so that
the routine can know the name of the view that needs to be dropped.