Add test to assert SaferAlterFieldSetNotNull error
The SaferAlterFieldSetNotNull operation fails to run when
the field field is a models.ForeignKey

This is because in a Django model, the name of a Python field may be
different from the name of the corresponding table column.

One such example is the ForeignKey field, which receives an "_id"
suffix in the database column name.