Move info for removing a unique constraint to its own guide
Why? The current information on removing a unique constraint is
clustered inside the class definition of the SaferRemoveUniqueConstraint
operation.

It is not obvious from a user's PoV that they will find a "guide" and a
"how-to" inside the documentation for the definition of a class as this
is not a common place for such information to live.

This commit moves this information to a more free-form "guides" page so
that the documentation is easier to find and easier to edit.

The class definition will still include a link back to the guide in
case users come to the class definition before they read the guide.