Run coverage on tests folder
Similar to the src/ folder, the tests/ folder can also have dead paths
or misconfiguration in test functions or fixtures that may lead to dead
tests, or false assumptions about what is being tested.

This change will start looking for lack of coverage on files in the
tests/ folder, and should help with finding out what these tests are in
the future.