Address lack of coverage in test_timeouts.py
This commit addresses the lines where codecoverage didn't hit in the
test_timeouts.py file.

The _Leaky class is designed to never hit certain code paths, and
because of that codecoverage will complain. This is expected, however,
and codecoverage can be ignored in these cases.

As for tests containing context managers like this:

  with pytest.raises(...):
      with timeouts.apply_timeouts(
          using="default",
          lock_timeout=lock_timeout,
          statement_timeout=statement_timeout,
      ):
          # This line will never be hit, because the context manager
          # raises before it can run.
          pass

An alternative would be to call __enter__() manually:

  timeouts.apply_timeouts(
      using="default",
      lock_timeout=lock_timeout,
      statement_timeout=statement_timeout,
  ).__enter__()

But that looks a bit unergonomic, and isn't the way this is used in
practice. Instead, a # pragma: no cover was slapped on it.