Bugfix: drop unique constraint sqlmigrate is noop
Change `collect_default` on constraint existing checks to require an
explicit value rather than defaulting to false. This was causing the operation
to say that the constraint did not exist and could not generate any SQL
when the migration to drop a constraint was in collect only mode.

The value needs to be set to True when dropping constraints so that the SQL
for what would happen if a constraint exists is shown (as the alternative is
not useful).