Add multi-schema behavioral tests for constraint queries
These tests set up two schemas containing identical tables and
constraints, in order to check that operations correctly check and act
only on the current schema (the one set in Postgres' search path).

These tests reveal that the operations *do not* act on the correct schema
reliably, which is a problem as Kraken gains multi-tenancy features and
will increasingly have multiple, largely-identical schemas (one for each
tenant). The tests have been marked with `xfail` and the operations will
be fixed in the next commit.

Related issue: #92