Fix: Fix to_field being ignored for SaferAddFieldForeignKey migrations
The `to_field` parameter can be explicitly provided to the django `ForeignKey` interface, however we
were defaulting to always use the pk of the related model. This led to a regression where django
would think the ForeignKey relationship was operating over the specified `to_field`, but the
database column type was incorrect.