Allow db only remove foreign key to work as expected
Remove foreign key for database was previously unable to apply
as it could not find the name of the field within the state.
This is modified to be able to function with only a name
and model provided, even if it is not on the state by handling
the error and allowing the ForeignKeyManager to be populated
with an optional field value.

It is noted that the Field value is still required in order for
other operations to occur however.